requirements.txt
numpy~=1.23.3
matplotlib~=3.5.3
pandas~=1.4.4
scikit-learn~=1.1.3
graphviz
mlrose_hiive
pymdptoolbox
sklearn~=0.0
gym~=0.26.2
scipy~=1.9.3